DoubleArray: Reduce size from 32 to 16 bits
All entries fit in 16 bits (but not 8).

PiperOrigin-RevId: 795345842

+++ src/base/strings/internal/double_array.h
@@ -40,8 +40,8 @@
 namespace mozc::japanese::internal {
 
 struct DoubleArray {
-  int32_t base;
-  uint32_t check;
+  int16_t base;
+  uint16_t check;
 };
 
 std::string ConvertUsingDoubleArray(const DoubleArray *da, const char *table,
